Add RepositoryUrl to FigmaPareNet.csproj
Added a `RepositoryUrl` element to the project file `FigmaPareNet.csproj`. This element specifies the URL of the repository where the source code is hosted, which is `https://github.com/xiaoshenl/FigmaPareNet`. This addition helps users and developers to easily find the source code and contribute to the project if needed.